# src/main/java/com/youssef/api-spec.json
{"openapi":"3.1.0","info":{"title":"Event Check-In API","description":"Event registration and race-safe door check-in","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/users":{"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/events":{"get":{"tags":["event-controller"],"operationId":"getAll","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageEventResponse"}}}}}},"post":{"tags":["event-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/v1/events/{id}/publish":{"post":{"tags":["event-controller"],"operationId":"publish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/v1/events/{eventId}/registrations":{"get":{"tags":["registration-controller"],"operationId":"getEventRegistrations","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRegistrationResponse"}}}}}},"post":{"tags":["registration-controller"],"operationId":"register","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegistrationResponse"}}}}}}},"/api/v1/events/{eventId}/check-ins":{"get":{"tags":["check-in-controller"],"operationId":"getAllCheckIns","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCheckInResponse"}}}}}},"post":{"tags":["check-in-controller"],"operationId":"checkIn","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCheckInRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckInResponse"}}}}}}},"/api/v1/attendees":{"get":{"tags":["attendee-controller"],"operationId":"getAll_1","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAttendeeResponse"}}}}}},"post":{"tags":["attendee-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAttendeeRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttendeeResponse"}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/registrations/{id}":{"get":{"tags":["registration-controller"],"operationId":"getRegistration","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegistrationResponse"}}}}}},"delete":{"tags":["registration-controller"],"operationId":"cancel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/events/{id}":{"get":{"tags":["event-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/v1/events/{eventId}/stats":{"get":{"tags":["event-controller"],"operationId":"getStats","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventStatsResponse"}}}}}}},"/api/v1/attendees/{id}":{"get":{"tags":["attendee-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttendeeResponse"}}}}}}},"/api/v1/events/{eventId}/delete":{"delete":{"tags":["event-controller"],"operationId":"deleteById","parameters":[],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8},"fullName":{"type":"string","maxLength":150,"minLength":0},"role":{"type":"string","enum":["ADMIN","ORGANIZER","STAFF"]}},"required":["email","fullName","password","role"]},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string","enum":["ADMIN","ORGANIZER","STAFF"]},"createdAt":{"type":"string","format":"date-time"}}},"CreateEventRequest":{"type":"object","properties":{"organizerId":{"type":"string","format":"uuid"},"name":{"type":"string","maxLength":200,"minLength":0},"venue":{"type":"string","maxLength":255,"minLength":0},"description":{"type":"string"},"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"},"checkInOpensAt":{"type":"string","format":"date-time"},"checkInClosesAt":{"type":"string","format":"date-time"},"capacity":{"type":"integer","format":"int32"}},"required":["capacity","checkInClosesAt","checkInOpensAt","endsAt","name","organizerId","startsAt"]},"EventResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizerId":{"type":"string","format":"uuid"},"organizerName":{"type":"string"},"name":{"type":"string"},"venue":{"type":"string"},"description":{"type":"string"},"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"},"checkInOpensAt":{"type":"string","format":"date-time"},"checkInClosesAt":{"type":"string","format":"date-time"},"capacity":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["DRAFT","PUBLISHED","CANCELLED","COMPLETED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateRegistrationRequest":{"type":"object","properties":{"attendeeId":{"type":"string","format":"uuid"}},"required":["attendeeId"]},"RegistrationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"eventId":{"type":"string","format":"uuid"},"attendeeId":{"type":"string","format":"uuid"},"eventName":{"type":"string"},"attendeeName":{"type":"string"},"status":{"type":"string","enum":["CONFIRMED","CANCELLED","WAITLISTED"]},"registeredAt":{"type":"string","format":"date-time"},"cancelledAt":{"type":"string","format":"date-time"},"ticketCode":{"type":"string"}}},"CreateCheckInRequest":{"type":"object","properties":{"ticketCode":{"type":"string"},"gate":{"type":"string"},"staffUserId":{"type":"string","format":"uuid"}},"required":["staffUserId","ticketCode"]},"CheckInResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"ticketCode":{"type":"string"},"checkedInBy":{"type":"string","format":"uuid"},"checkedInAt":{"type":"string","format":"date-time"},"gate":{"type":"string"}}},"CreateAttendeeRequest":{"type":"object","properties":{"email":{"type":"string"},"fullName":{"type":"string","maxLength":150,"minLength":0},"phone":{"type":"string"}},"required":["email","fullName"]},"AttendeeResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"fullName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageEventResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/EventResponse"}},"number":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}},"EventStatsResponse":{"type":"object","properties":{"eventId":{"type":"string","format":"uuid"},"eventName":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"confirmedRegistrations":{"type":"integer","format":"int64"},"checkedIn":{"type":"integer","format":"int64"},"notYetArrived":{"type":"integer","format":"int64"},"attendanceRate":{"type":"number","format":"double"}}},"PageRegistrationResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RegistrationResponse"}},"number":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageCheckInResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CheckInResponse"}},"number":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageAttendeeResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AttendeeResponse"}},"number":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}